+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - New `step` argument to `rics.ml.time_split` splitting functions.
+- New argument `attrs` to the `misc.name()`-function. Default is `'func'` (name used by `functools.partial`).
 
 ## [3.1.0] - 2023-10-13